factor out path diffing logic
