+++ compiler/crates/relay-codegen/Cargo.toml
@@ -43,7 +43,6 @@ md-5 = "0.8"
 relay-config = { path = "../relay-config" }
 relay-transforms = { path = "../relay-transforms" }
 schema = { path = "../schema" }
-pathdiff = "0.2.1"
 
 [dev-dependencies]
 fixture-tests = { path = "../fixture-tests" }
