+++ compiler/crates/relay-config/Cargo.toml
@@ -16,3 +16,4 @@ serde = { version = "1.0.136", features = ["derive", "rc"] }
 serde_json = { version = "1.0.79", features = ["float_roundtrip", "unbounded_depth"] }
 strum = "0.21"
 strum_macros = "0.21"
+pathdiff = "0.2.1"
