+++ compiler/crates/relay-typegen/Cargo.toml
@@ -21,7 +21,6 @@ lazy_static = "1.0"
 relay-config = { path = "../relay-config" }
 relay-transforms = { path = "../relay-transforms" }
 schema = { path = "../schema" }
-pathdiff = "0.2.1"
 
 [dev-dependencies]
 fixture-tests = { path = "../fixture-tests" }
